Update GPAcal.py
Basically, I just tried accounting for possible errors on the part of the user so instead of the code sopping, it just alerts the user to correct himself. But then, I could only account for just a chance for a trial. If similar error occur again, which has a lesser chance of happening, the traceback error displays again. There's still more improvement though. Happy learning Tife